+++ pytest.ini
@@ -4,7 +4,7 @@ testpaths = gdsfactory/
 addopts = '--tb=short'
 python_files = gdsfactory/*.py notebooks/*.ipynb
 collect_ignore = "gdsfactory/difftest.py"
-norecursedirs = gdsfactory/simulation/gmeep gdsfactory/simulation/modes gdsfactory/simulation/gtidy3d
+norecursedirs = gdsfactory/simulation/gmeep gdsfactory/simulation/modes gdsfactory/simulation/tidy3d
 
 ; ignore = 'difftest.py'
 ; nb_file_fnmatch = notebooks/*.ipynb
